missing declaration and size_t vs int usage

+++ redis.c
@@ -3063,9 +3063,9 @@ PHP_METHOD(Redis, pubsub) {
 
 // Construct an EVAL or EVALSHA command, with option argument array and number
 // of arguments that are keys parameter
-PHP_REDIS_API int
+PHP_REDIS_API size_t
 redis_build_eval_cmd(RedisSock *redis_sock, char **ret, char *keyword,
-                     char *value, int val_len, zval *args, int keys_count
+                     char *value, size_t val_len, zval *args, int keys_count
                      TSRMLS_DC)
 {
     zval *elem;
@@ -3214,7 +3214,7 @@ PHP_METHOD(Redis, eval)
 
     // Construct our EVAL command
     cmd_len = redis_build_eval_cmd(redis_sock, &cmd, "EVAL", script, script_len,
-        args, keys_count TSRMLS_CC);
+        args, (int)keys_count TSRMLS_CC);
 
     REDIS_PROCESS_REQUEST(redis_sock, cmd, cmd_len);
     IF_ATOMIC() {
